move test_non_reducing_slice_on_multiindex
