+++ pandas/tests/indexing/test_indexing.py
@@ -812,23 +812,6 @@ class TestMisc(Base):
             tslice_ = _non_reducing_slice(slice_)
             assert isinstance(df.loc[tslice_], DataFrame)
 
-    def test_non_reducing_slice_on_multiindex(self):
-        # GH 19861
-        dic = {
-            ('a', 'd'): [1, 4],
-            ('a', 'c'): [2, 3],
-            ('b', 'c'): [3, 2],
-            ('b', 'd'): [4, 1]
-        }
-        df = pd.DataFrame(dic, index=[0, 1])
-        idx = pd.IndexSlice
-        slice_ = idx[:, idx['b', 'd']]
-        tslice_ = _non_reducing_slice(slice_)
-
-        result = df.loc[tslice_]
-        expected = pd.DataFrame({('b', 'd'): [4, 1]})
-        tm.assert_frame_equal(result, expected)
-
     def test_list_slice(self):
         # like dataframe getitem
         slices = [['A'], Series(['A']), np.array(['A'])]
